shared/zipContext and utils in one file
